Break p2p_client.integration.test into smaller files
