+++ .test_patterns.yml
@@ -212,7 +212,7 @@ tests:
     error_regex: "✕ should stop after max retry attempts"
     owners:
       - *sean
-  - regex: "p2p/src/client/p2p_client.integration.test.ts" # http://ci.aztec-labs.com/31c4c9821b1a533c
+  - regex: "p2p/src/client/p2p_client/test/" 
     owners:
       - *palla
   - regex: "yarn-project/kv-store"
